Module google_cloudprofiler2::api

Structs

CloudProfiler
Central instance to access all CloudProfiler related resource activities

CreateProfileRequest
CreateProfileRequest describes a profile resource online creation request. The deployment field must be populated. The profile_type specifies the list of profile types supported by the agent. The creation call will hang until a profile of one of these types needs to be collected.

Deployment
Deployment contains the deployment identification information.

Profile
Profile resource.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudProfiler hub.

ProjectProfileCreateCall
CreateProfile creates a new profile resource in the online mode. The server ensures that the new profiles are created at a constant rate per deployment, so the creation request may hang for some time until the next profile session is available. The request may fail with ABORTED error if the creation is not available within ~1m, the response will indicate the duration of the backoff the client should take before attempting creating a profile again. The backoff duration is returned in google.rpc.RetryInfo extension on the response status. To a gRPC client, the extension will be return as a binary-serialized proto in the trailing metadata item named "google.rpc.retryinfo-bin".

ProjectProfileCreateOfflineCall
CreateOfflineProfile creates a new profile resource in the offline mode. The client provides the profile to create along with the profile bytes, the server records it.

ProjectProfilePatchCall
UpdateProfile updates the profile bytes and labels on the profile resource created in the online mode. Updating the bytes for profiles created in the offline mode is currently not supported: the profile content must be provided at the time of the profile creation.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.